refactor cache code
